Extract Assessors to common interface
