+++ sonar-javascript-plugin/src/main/java/org/sonar/plugins/javascript/filter/MinificationAssessor.java
@@ -31,7 +31,7 @@ import org.sonar.api.utils.log.Loggers;
  * This situation is typical of files that have been generated by some tool.
  * Such files are of poor interest as regards a SonarQube analysis.
  */
-public class MinificationAssessor {
+class MinificationAssessor implements Assessor {
 
   private static final Logger LOG = Loggers.get(MinificationAssessor.class);
 
@@ -72,4 +72,8 @@ public class MinificationAssessor {
     return averageLineLength > averageLineLengthThreshold;
   }
 
+  @Override
+  public boolean test(InputFile inputFile) {
+    return isMinified(inputFile);
+  }
 }
